TP1/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.11
Flask==3.1.0
matplotlib==3.9.0
requests==2.32.3
